+++ src/vanished/Simulator/Structure/StockManagerInfo.java
@@ -6,12 +6,12 @@ import vanished.Simulator.Item.ItemDef;
 
 public class StockManagerInfo {
 	public ItemDef itemDef;
-	public int lotmax;
-	public int capacity;
+	public double lotmax;
+	public double capacity;
 
 	public StockManagerInfo(ItemDef itemDef, String prefix, Properties p) {
 		this.itemDef = itemDef;
-		lotmax = Integer.parseInt(p.getProperty(prefix + "lotmax"));
-		capacity = Integer.parseInt(p.getProperty(prefix + "capacity"));
+		lotmax = Double.parseDouble(p.getProperty(prefix + "lotmax"));
+		capacity = Double.parseDouble(p.getProperty(prefix + "capacity"));
 	}
 }
